convert stave fix for lat long being made to be integers

+++ R/convert_stave.R
@@ -35,8 +35,8 @@ convert_stave <- function(data) {
                   latitude = .data$lat, longitude = .data$lon, .data$spatial_notes, 
                   .data$collection_start, .data$collection_end, 
                   .data$collection_day, .data$time_notes) %>%
-    dplyr::mutate(latitude = as.integer(.data$latitude)) %>% 
-    dplyr::mutate(longitude = as.integer(.data$longitude)) %>% 
+    dplyr::mutate(latitude = as.numeric(.data$latitude)) %>% 
+    dplyr::mutate(longitude = as.numeric(.data$longitude)) %>% 
     dplyr::distinct(.data$study_key, .data$survey_ID, .keep_all = TRUE) %>% 
     dplyr::rename(survey_id = survey_ID)
